% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{full.factorial}
\alias{full.factorial}
\title{Full factorial design.}
\usage{
full.factorial(p, level)
}
\arguments{
\item{p}{design dimension.}

\item{level}{an integer specifying the number of levels.}
}
\value{
a full factorial design matrix (scaled to [0, 1])
}
\description{
This function generates a full factorial design.
}
\details{
\code{full.factorial} generates a p dimensional full factorial design.
}
\examples{
p = 3
level = 3
D = full.factorial(p, level)

}
