% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{randomLHD}
\alias{randomLHD}
\title{Random Latin hypercube design}
\usage{
randomLHD(n, p)
}
\arguments{
\item{n}{design size.}

\item{p}{design dimension.}
}
\value{
a random Latin hypercube design.
}
\description{
This function generates a random Latin hypercube design.
}
\details{
\code{randomLHD} generates a random Latin hypercube design.
}
\examples{
n = 20
p = 3
D = randomLHD(n, p)

}
