\name{SFDesign-package}
\alias{SFDesign-package}
\docType{package}
\title{
  Space-Filling designs.
}
\description{
This pacakge offers a comprehensive suite of functions to construct various types of space-filling designs, including Latin hypercube designs, clustering-based designs, maximin designs, maximum projection designs, and uniform designs (Joseph 2016). It also offers the option to optimize designs based on user-defined criteria.
}
\author{
Shangkun Wang, V. Roshan Joseph

Maintainer: Shangkun Wang <shangkunwang01@gmail.com>
}
\references{
Wang, Shangkun, Xie, Weijun and V. Roshan Joseph. SFDesign: An R package for Space-Filling Designs.

Joseph, V. R. (2016). Space-filling designs for computer experiments: A review. Quality Engineering, 28(1), 28-35.
}
